move option to content map, update lib
