+++ package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@tanstack/react-router": "^1.112.0",
-    "mjqt-scoring": "^0.1.20",
+    "mjqt-scoring": "^0.1.22",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
